[Data] Separate test_auto_batch_size.py with integration and unit tests (#62874)
## Description

Splits `test_auto_batch_size.py` into unit and integration test files
following the `tests/unit/` convention.

Moved to `tests/unit/` (5 tests):

* `test_compute_auto_batch_size_basic`
* `test_compute_auto_batch_size_clamped_to_one`
* `test_compute_auto_batch_size_returns_none` (parametrized:
`empty_iterator`, `zero_rows`)
* `test_compute_auto_batch_size_iterator_includes_peeked_block`
* `test_auto_batches_respect_target_size`

Remain in integration file (1 test):

* `test_map_batches_auto_correctness` (uses `ray_start_regular_shared`,
`ray.data`)

## Related issues

Related to #61125

## Additional information

Tests were classified by whether they use `ray_start_*` fixtures or make
runtime `ray.*` calls — not by import paths, since the unit tests still
import from `ray.data._internal.*` to exercise internal classes
directly.

---------
